% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportionDaysCovered.R
\name{addHospital}
\alias{addHospital}
\title{Adds hospital days}
\usage{
addHospital(treatmentName, treatmentTable, myHospitalDays)
}
\arguments{
\item{treatmentName}{the row ID that defines the patient and drug class}

\item{treatmentTable}{the complete treatment table}

\item{myHospitalDays}{the hospital days for all patients}
}
\value{
hopsital days added to the treatment table
}
\description{
Adds hospital days
}
