% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportionDaysCovered.R
\name{addUncovered}
\alias{addUncovered}
\title{Adds uncovered and breaks days into the treatment of one patient and one drug class}
\usage{
addUncovered(treatmentName, treatmentTable, myBreakDays)
}
\arguments{
\item{treatmentName}{the row ID that defines the patient and drug class}

\item{treatmentTable}{the complete treatment table}

\item{myBreakDays}{the break days for all drug classes}
}
\value{
uncovered and breaks days for one patient and one drug class
}
\description{
Adds uncovered and breaks days into the treatment of one patient and one drug class
}
