% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responseScore.R
\name{adjustScore}
\alias{adjustScore}
\title{Get the score from the first PC}
\usage{
adjustScore(
  responseDf,
  scoreColumn = "phenoScore",
  confounderColumns,
  responseColumns
)
}
\arguments{
\item{responseDf}{the data frame that contains the time blocks and the scores and the confounders per time block}

\item{scoreColumn}{name of the column that contains the score}

\item{confounderColumns}{names of the columns that contain confounding information}

\item{responseColumn}{names of the columns containing the patient block value for a biomarker (e.g. vector of string names). The score will be based only on these columns.}
}
\value{
the response data frame with an adjusted score
}
\description{
Get the score from the first PC
}
\keyword{internal}
