% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportionDaysCovered.R
\name{calculateAdherences}
\alias{calculateAdherences}
\title{Controls the adherence calculation for the whole treatment table}
\usage{
calculateAdherences(treatmentTable, startDates, endDates, refillPeriod)
}
\arguments{
\item{treatmentTable}{the complete treatment table matrix}

\item{startDates}{a data frame containing the start dates of the study for each patient}

\item{endDates}{a data frame containing the end dates of the study for each patient}

\item{refillPeriod}{length of a prescription refill period in days}
}
\value{
adherences for the entire treatment table as a data frame
}
\description{
Controls the adherence calculation for the whole treatment table
}
