% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFixedAdherence.R
\name{calculateFixedAdherenceFixedPeriod}
\alias{calculateFixedAdherenceFixedPeriod}
\title{Calculates medication adherence between start and end dates}
\usage{
calculateFixedAdherenceFixedPeriod(
  serialDf,
  startDates,
  endDates,
  atcCode = "C09",
  refillPeriod = 90,
  idColumn = "PATIENT",
  dateColumn = "VISIT",
  atcColumn = "ATC"
)
}
\arguments{
\item{serialDf}{the data frame with the serial data}

\item{startDates}{dataframe with ID and a date to start the adherence calculation}

\item{endDates}{dataframe with ID and a date to end the adherence calculation}

\item{atcCode}{regular expression for the class of medication ATC code (default: C09 which stands for any RAAS)}

\item{refillPeriod}{length of the refill period (default 90 days)}

\item{idColumn}{name of ID column: default is PATIENT}

\item{dateColumn}{name of date column: default is VISIT. This column has to be of class Date}

\item{atcColumn}{name of the column with the ATC codes: default is ATC}
}
\value{
time block data and dates in wide format
}
\description{
Calculates medication adherence between start and end dates
}
