% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFixedAdherence.R
\name{calculateFixedAdherenceList}
\alias{calculateFixedAdherenceList}
\title{Calculates medication adherence based on a fixed, specified refill time period}
\usage{
calculateFixedAdherenceList(
  serialDf,
  startDates = NA,
  endDates = NA,
  atcCodeList = c("C09", "C10"),
  refillPeriod = 90,
  stopPeriod = NA,
  idColumn = "PATIENT",
  dateColumn = "VISIT",
  atcColumn = "ATC",
  doPCAScore = F,
  combinedScore = T
)
}
\arguments{
\item{serialDf}{the data frame with the serial data}

\item{startDates}{dataframe with ID and a date to start the adherence calculation (optional)}

\item{endDates}{dataframe with ID and a date to end the adherence calculation (optional)}

\item{refillPeriod}{length of the refill period (default 90 days)}

\item{stopPeriod}{number of days after which the medication is considered to have stopped (optional)}

\item{idColumn}{name of ID column: default is PATIENT}

\item{dateColumn}{name of date column: default is VISIT. This column has to be of class Date}

\item{atcColumn}{name of the column with the ATC codes: default is ATC}

\item{doPCAScore}{flag whether a PCA score from the medication data should be calculated}

\item{combinedScore}{flag whether a combined PCA scores for all ATC codes in the list should be calculated}

\item{atcCode}{list of regular expressions for the classes of medication ATC code (default: c('C09', 'C10'))}
}
\value{
time block data and dates in wide format
}
\description{
Calculates medication adherence based on a fixed, specified refill time period
}
