% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{calculateInterpolation}
\alias{calculateInterpolation}
\title{calculate the linear interpolation from one measurement to the next available one}
\usage{
calculateInterpolation(toInterpolate, dataDF, dateDF, timeframe,
  baselineDate)
}
\arguments{
\item{toInterpolate}{index position of the start of the interpolation}

\item{dataDF}{the data frame with the measurements records}

\item{dateDF}{corresponding data frame with date records}

\item{timeframe}{length of blocks in days}

\item{baselineDate}{the baseline date}
}
\value{
data frame with (linearly) interpolated values
}
\description{
calculate the linear interpolation from one measurement to the next available one
}
\keyword{internal}
