% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportionDaysCovered.R
\name{calculateMedicationSpecificAdherence}
\alias{calculateMedicationSpecificAdherence}
\title{Calculate the adherence for each patient and all their drug classes}
\usage{
calculateMedicationSpecificAdherence(
  treatmentRowNames,
  treatment,
  startDates,
  endDates,
  refillPeriod
)
}
\arguments{
\item{treatmentRowNames}{the row ID from the treatment table identifying one patent and one drug class}

\item{treatment}{the complete treatment table matrix}

\item{startDates}{a data frame containing the start dates of the study for each patient}

\item{endDates}{a data frame containing the end dates of the study for each patient}

\item{refillPeriod}{length of a prescription refill period in days}
}
\value{
adherences for a patient and drug class
}
\description{
Calculate the adherence for each patient and all their drug classes
}
