% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportionDaysCovered.R
\name{calculateOneAdherenceValue}
\alias{calculateOneAdherenceValue}
\title{Calculate one adherence rate}
\usage{
calculateOneAdherenceValue(treatmentRow, refillPeriod)
}
\arguments{
\item{treatmentRow}{one row from the treatment table identifying one patient and one drug class}

\item{refillPeriod}{length of a prescription refill period in days}
}
\value{
adherence
}
\description{
Calculate one adherence rate
}
