% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responseScore.R
\name{calculateScore}
\alias{calculateScore}
\title{Get the score from the first PC}
\usage{
calculateScore(responseDf, responseColumns, firstPrincipalComponent)
}
\arguments{
\item{responseDf}{the data frame that contains the time blocks and the biomarker data per time block}

\item{firstPrincipalComponent}{a vector containing the first PC}

\item{responseColumn}{names of the columns containing the patient block value for a biomarker (e.g. vector of string names). The score will be based only on these columns.}
}
\value{
the response data frame with an added column phenoScore that shows the computed score
}
\description{
Get the score from the first PC
}
\keyword{internal}
