% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{checkBaselineFormat}
\alias{checkBaselineFormat}
\title{Checks the format of the data set for the baseline dates}
\usage{
checkBaselineFormat(df, idColumn = "PATIENT", dateColumn = "VISIT")
}
\arguments{
\item{df}{the data frame to check}

\item{idColumn}{name of ID column: default is PATIENT}

\item{dateColumn}{name of date column: default is VISIT. This column has to be of class Date}
}
\value{
NA if there are format problems, otherwise a data frame that has a PATIENT andVISIT column
}
\description{
We need an ID and a date column. the date column has to be of class Date.
}
\keyword{internal}
