% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{checkBaselineHospitalFormat}
\alias{checkBaselineHospitalFormat}
\title{Checks the format of the hospital data}
\usage{
checkBaselineHospitalFormat(df, idColumn = "PATIENT",
  startDateColumn = "VISIT_START", endDateColumn = "VISIT_END")
}
\arguments{
\item{df}{the data frame to check}

\item{idColumn}{name of ID column: default is PATIENT}

\item{startDateColumn}{name of date column: default is VISIT_START. This column has to be of class Date}

\item{endDateColumn}{name of date column: default is VISIT_END. This column has to be of class Date}
}
\value{
NULL if there are format problems, otherwise a data frame that has a PATIENT andVISIT column
}
\description{
We need an ID and two date columns. The date columns have to be of class Date.
}
\keyword{internal}
