% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{checkBinaryPrescriptionFormat}
\alias{checkBinaryPrescriptionFormat}
\title{Checks the format of the data set for the prescriptions (without DDDs)}
\usage{
checkBinaryPrescriptionFormat(
  df,
  idColumn = "PATIENT",
  dateColumn = "VISIT",
  atcColumn = "ATC"
)
}
\arguments{
\item{df}{the data frame to check}

\item{idColumn}{name of ID column: default is PATIENT}

\item{dateColumn}{name of date column: default is VISIT. This column has to be of class Date}

\item{atcColumn}{name of the ATC column: default is ATC}
}
\value{
NULL if there are format problems, otherwise a data frame that has a PATIENT, VISIT, ATC
}
\description{
We need an ID, date column and ATC. The date column has to be of class Date.
}
\keyword{internal}
