% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{checkBlockFormat}
\alias{checkBlockFormat}
\title{Checks the format of the data set for the time block building}
\usage{
checkBlockFormat(
  df,
  idColumn = "PATIENT",
  dateColumn = "VISIT",
  dataColumn = "MEASURE"
)
}
\arguments{
\item{df}{the data frame to check}

\item{idColumn}{name of ID column: default is PATIENT}

\item{dateColumn}{name of date column: default is VISIT. This column has to be of class Date}

\item{dataColumn}{name of the data column: default is MEASURE}
}
\value{
NULL if there are format problems, otherwise a data frame that has a PATIENT, VISIT and MEASURE column
}
\description{
We need an ID, data and a date column. the date column has to be of class Date.
}
\keyword{internal}
