% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{checkStructureColumnResponse}
\alias{checkStructureColumnResponse}
\title{Checks whether a column is in a data frame}
\usage{
checkStructureColumnResponse(columnName, df)
}
\arguments{
\item{columnName}{name of the column}

\item{df}{the data frame to check}
}
\value{
TRUE if the  column passes the test
}
\description{
Checks whether a column is in a data frame
}
\keyword{internal}
