% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{checkVizFormat}
\alias{checkVizFormat}
\title{Checks the format of the data set quick visualisation}
\usage{
checkVizFormat(
  df,
  idColumn = "PATIENT",
  blockColumn = "block",
  dataColumn = "measure",
  dateColumn = "date.measure",
  prescriptionColumn = "prescription"
)
}
\arguments{
\item{df}{the response data frame. It needs to contain binary prescription information.}

\item{idColumn}{name of ID column: default is PATIENT}

\item{blockColumn}{name of the block column: default is block}

\item{dataColumn}{name of the data column: default is measure}

\item{dateColumn}{name of date column: default is date.measure. This column has to be of class Date}

\item{prescriptionColumn}{name of the prescription column: default is prescription (it has to be binary)}
}
\value{
NULL if there are format problems, otherwise a data frame that has a PATIENT, VISIT, MEASURE, BLOCK and PRESCRIPTION column
}
\description{
Checks the format of the data set quick visualisation
}
\keyword{internal}
