% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportionDaysCovered.R
\name{combineRows}
\alias{combineRows}
\title{Combine the information from all drug classes into one row (logical AND)}
\usage{
combineRows(myValues)
}
\arguments{
\item{myValue}{0, 1, 2 or NA from different drug classes on the same date}
}
\value{
combined row
}
\description{
Combine the information from all drug classes into one row (logical AND)
}
