% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHospitalisationTimeBlocks.R
\name{createHospitalTimeBlocks}
\alias{createHospitalTimeBlocks}
\title{Calculates block data for hospitalisations}
\usage{
createHospitalTimeBlocks(serialDf, baselineInfo, studyStartDate,
  studyEndDate, blockDays = 365.25, blDayDiff = 90,
  idColumn = "PATIENT", startDateColumn = "VISIT_START",
  endDateColumn = "VISIT_END", countNights = TRUE,
  blDateColumn = "VISIT", longFormat = TRUE)
}
\arguments{
\item{serialDf}{the data frame with the hospital events (one row per event)}

\item{baselineInfo}{dataframe with ID and baseline date}

\item{studyStartDate}{start Date of the study (has to be of class Date)}

\item{studyEndDate}{end Date of the study (has to be of class Date)}

\item{blockDays}{length of a time block in days (default 365.25 days)}

\item{blDayDiff}{time in days that should be searched for a measurement around the baseline visit (default 90)}

\item{idColumn}{name of ID column: default is PATIENT}

\item{startDateColumn}{name of start date column (day of hospitalisation or outpatient visit): default is VISIT_START. This column has to be of class Date}

\item{endDateColumn}{name of end date column (release date): default is VISIT_END. This column has to be of class Date}

\item{countNights}{flag indicating whether to count the days in hospital or the night: default TRUE, count nights (if nights are counted, an outpatient stay has length 0, otherwise 1)}

\item{blDateColumn}{name of the column containing the baseline date (date has to be of class Date): default is VISIT}

\item{longFormat}{flag to indicate whether results should be in long or wide format. Wide has some additional columns (default: TRUE)}
}
\value{
time block data and dates in wide format
}
\description{
Calculates block data for hospitalisations
}
