% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMedicationTimeBlocks.R
\name{createMedicationTimeBlocks}
\alias{createMedicationTimeBlocks}
\title{Calculates the time block data for prescriptions}
\usage{
createMedicationTimeBlocks(
  serialDf,
  baselineInfo,
  studyStartDate,
  studyEndDate,
  atcCode = "C09",
  blockDays = 365.25,
  useDoses = FALSE,
  idColumn = "PATIENT",
  dateColumn = "VISIT",
  atcColumn = "ATC",
  dddColumn = "cDDD",
  dataName = "prescription",
  longFormat = TRUE,
  noUseAtBaseline = FALSE
)
}
\arguments{
\item{serialDf}{the data frame with the serial data}

\item{baselineInfo}{dataframe with ID and baseline date}

\item{studyStartDate}{start Date of the study (has to be of class Date)}

\item{studyEndDate}{end Date of the study (has to be of class Date)}

\item{atcCode}{regular expression for the class of medication ATC code (default: C09 which stands for any RAAS)}

\item{blockDays}{length of a time block in days (default 365.25 days)}

\item{useDoses}{flag indicating whether presence of prescription events or sum of doses should be used in a block (default FALSE)}

\item{idColumn}{name of ID column: default is PATIENT}

\item{dateColumn}{name of date column: default is VISIT. This column has to be of class Date}

\item{atcColumn}{name of the column with the ATC codes: default is ATC}

\item{dddColumn}{name of the column with the dosage information: default is cDDD}

\item{dataName}{set a user specified name for the data column in the output: default is prescription}
}
\value{
time block data and dates in wide format
}
\description{
Calculates the time block data for prescriptions
}
\examples{
\dontrun{
baselinePatients <- read_tsv('../data/ourBaselinePatients.txt', guess_max = 3582) \%>\%
    select(PATIENT, VISIT = VISIT_TL)
allFimea <- read_tsv('../data/all_fimea_kela.txt')
allFimea <- allFimea \%>\% mutate(VISIT = ymd(allFimea$ostopv))
allFimea <- allFimea \%>\% mutate(PATIENT = fd)
allFimea <- allFimea \%>\% filter(!is.na(all_ddd)) #check these as they should not exists
medResLong <- createMedicationTimeBlocks(allFimea, baselinePatients,
   studyStartDate = as.Date('1995-01-01', '\%Y-\%m-\%d'),
   studyEndDate = as.Date('2015-12-31', '\%Y-\%m-\%d'),
   atcCode = 'C09', blockDays = 365.25, useDoses = FALSE, idColumn = "PATIENT", dateColumn = "VISIT",
   atcColumn = "ATC_FIMEA", dddColumn = 'all_ddd', dataName = "anyRAAS", longFormat = TRUE,
   noUseAtBaseline = FALSE)
medResWide <- createMedicationTimeBlocks(allFimea, baselinePatients,
   studyStartDate = as.Date('1995-01-01', '\%Y-\%m-\%d'),
   studyEndDate = as.Date('2015-12-31', '\%Y-\%m-\%d'),
   atcCode = 'C09', blockDays = 365.25, useDoses = FALSE, idColumn = "PATIENT", dateColumn = "VISIT",
   atcColumn = "ATC_FIMEA", dddColumn = 'all_ddd', dataName = "anyRAAS", longFormat = FALSE,
   noUseAtBaseline = FALSE)
}
}
