% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTimeBlocks.R
\name{createTimeBlocks}
\alias{createTimeBlocks}
\title{Calculates the time block data  in wide format}
\usage{
createTimeBlocks(serialDf, baselineInfo, studyStartDate, studyEndDate,
  blockDays = 365.25, interpolate = TRUE, measureIsFactor = FALSE,
  blockSummaryStats = "median", blDayDiff = 90, idColumn = "PATIENT",
  dateColumn = "VISIT", dataColumn = "MEASURE", dataName = "measure",
  longFormat = TRUE)
}
\arguments{
\item{serialDf}{the data frame with the serial data}

\item{baselineInfo}{dataframe with ID and baseline date}

\item{studyStartDate}{start Date of the study (has to be of class Date)}

\item{studyEndDate}{end Date of the study (has to be of class Date)}

\item{blockDays}{length of a time block in days (default 365.25 days)}

\item{interpolate}{shall data for empty blocks be interpolated (default TRUE)}

\item{measureIsFactor}{when FALSE (default) data will be interpolated linearly, otherwise, empty blocks will take the value of the last preceding non-empty block}

\item{blockSummaryStats}{how to summarise the data in each block. Currently available: min, max, median, IQR}

\item{blDayDiff}{time in days that should be searched for a measurement around the baseline visit (default 90)}

\item{idColumn}{name of ID column: default is PATIENT}

\item{dateColumn}{name of date column: default is VISIT. This column has to be of class Date}

\item{dataColumn}{name of the data column: default is MEASURE}

\item{dataName}{set a user specified name for the data column in the output: default is measure}
}
\value{
time block data and dates in wide format
}
\description{
Calculates the time block data  in wide format
}
