% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{findBaselineHospitalStay}
\alias{findBaselineHospitalStay}
\title{Finds the length of a hospital stay at the baseline block}
\usage{
findBaselineHospitalStay(serialDf, baselineDates = baselineInfo,
  blDayDiff = blDayDiff, idColumn = "PATIENT",
  startDateColumn = "VISIT_START", endDateColumn = "VISIT_END",
  blDateColumn = "VISIT", countNights = TRUE)
}
\arguments{
\item{idColumn}{name of ID column: default is PATIENT}

\item{startDateColumn}{name of date column: default is VISIT_START. This column has to be of class Date}

\item{endDateColumn}{name of date column: default is VISIT_END. This column has to be of class Date}

\item{countNights}{whether to count nights or days in hopsital (if night, than an outpatient stay has length 0, otherwise 1)}

\item{df}{the data frame to check}

\item{dayDiff}{time around baseline (default: 90 days, will look three months before and after baseline)}
}
\value{
data frame with baseline measure or NA if no measurement was close to baseline
}
\description{
Finds the length of hospital stays around baseline
}
