% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{findClosestToBaseline}
\alias{findClosestToBaseline}
\title{Finds the measurement date that is closest to baseline}
\usage{
findClosestToBaseline(serialDf, baselineDates, blDayDiff = 90, ...)
}
\arguments{
\item{df}{the data frame to check}

\item{dayDiff}{time around baseline (default: 90 days, will look three months before and after baseline)}
}
\value{
data frame with baseline measure or NA if no measurement was close to baseline
}
\description{
Finds the measurement date that is closest to baseline in the given time period around baseline.
}
\keyword{internal}
