% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{findPrescriptionstAtBaseline}
\alias{findPrescriptionstAtBaseline}
\title{Finds the prescription information around the baseline visit}
\usage{
findPrescriptionstAtBaseline(prescriptions, baselineDates, anyPrescriptions,
  blockDays, useDoses)
}
\arguments{
\item{baselineDates}{the dataframe with all patient IDs and baseline visit dates}

\item{blockDays}{length of a block. Baseline info will be searched for the block of length blockDays where the baseline visit sits right in the middle}

\item{serialDf}{the data frame with all the prescription data}

\item{usesDoses}{flag indicating whether to use the presence of a prescription or the cumulate DDDs}
}
\value{
data frame with baseline prescription information
}
\description{
Finds the prescription information around the baseline visit by looking half a block on either side of baseline.
The information is either returned in binary form or in cumulative DDDs depending on the useDoses flag
}
\keyword{internal}
