% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFixedAdherence.R
\name{listAllAdherences}
\alias{listAllAdherences}
\title{Helper function to call individual (one ATC code) adherence calculation for a list of ATC codes}
\usage{
listAllAdherences(
  atcCode,
  serialDf,
  startDates,
  endDates,
  refillPeriod,
  stopPeriod,
  idColumn,
  dateColumn,
  atcColumn
)
}
\arguments{
\item{atcCode}{regular expression for the class of medication ATC code}

\item{serialDf}{the data frame with the serial data}

\item{startDates}{dataframe with ID and a date to start the adherence calculation (optional)}

\item{endDates}{dataframe with ID and a date to end the adherence calculation (optional)}

\item{refillPeriod}{length of the refill period (default 90 days)}

\item{stopPeriod}{number of days after which the medication is considered to have stopped (optional)}

\item{idColumn}{name of ID column: default is PATIENT}

\item{dateColumn}{name of date column: default is VISIT. This column has to be of class Date}

\item{atcColumn}{name of the column with the ATC codes: default is ATC}
}
\value{
time block data and dates in wide format
}
\description{
Helper function to call individual (one ATC code) adherence calculation for a list of ATC codes
}
