% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{makeLong}
\alias{makeLong}
\title{Transform the time block results from wide to long format}
\usage{
makeLong(df, dataName, idColumn)
}
\arguments{
\item{df}{the resulting wide format time block data}

\item{dataName}{the name of the data column in the output}

\item{idColumn}{the name of the ID column}
}
\value{
data frame in long format
}
\description{
Transform the time block results from wide to long format
}
\keyword{internal}
