% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{noUseAtMissingBaseline}
\alias{noUseAtMissingBaseline}
\title{Set missing medication info at baseline and immediately after to 0 which means no use}
\usage{
noUseAtMissingBaseline(df, blockLength, dataName)
}
\arguments{
\item{df}{wide format result data frame}

\item{blockLength}{length of consecutive blocks that are empty after baseline}

\item{dataName}{name of the data columns}
}
\value{
data frame with no medication use added at baseline and immediately after if missing
}
\description{
Use this function carefully as it assumes that missing data means no prescritpions
}
\keyword{internal}
