% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportionDaysCovered.R
\name{pdc_treatment}
\alias{pdc_treatment}
\title{Calculates adherence to drug treatments based on the proportion of days covered method.}
\usage{
pdc_treatment(
  serialDf,
  startDates,
  endDates,
  atcCode = c(),
  refillPeriod = 90,
  idColumn = "PATIENT",
  dateColumn = "VISIT",
  atcColumn = "ATC",
  treatmentBreakDays = c(),
  createGraphs = F
)
}
\arguments{
\item{serialDf}{the data frame with prescription data}

\item{startDates}{a data frame containing the start dates of the study for each patient}

\item{endDates}{a data frame containing the end dates of the study for each patient}

\item{atcCode}{a vector containing regular expressions, each encoding for one component/drug class of the treatment}

\item{refillPeriod}{length of a prescription refill period in days (default 90 days)}

\item{idColumn}{name of ID column: default is PATIENT}

\item{dateColumn}{name of date column: default is VISIT. This column has to be of class Date}

\item{atcColumn}{name of the column with the ATC codes: default is ATC}

\item{treatmentBreakDays}{a vector containing the number of days (one entry for each drug class) after which the treatment is considered discontinued (default: no breaks applied)}

\item{createGraphs}{flag indicating whether graphs should be produced (default: FALSE)}
}
\value{
adherence rates for the full prescription period and between start and end dates
}
\description{
Calculates adherence to drug treatments based on the proportion of days covered method.
}
\examples{
\dontrun{
 dfStart <- read_tsv('/home/ad/home/s/stefmutt/projects/former/cadGRS/data/bl_all_new.txt') \%>\%
   select(PATIENT, VISIT)
 dfEnd <- dfStart \%>\%
   mutate(VISIT = as.Date('2015-12-31'))
 
 kela <- read_tsv('/home/ad/home/s/stefmutt/projects/former/cadGRS/data/kela_all.txt') \%>\%
   filter(!is.na(ATC)) \%>\%
   select(-all_ddd) \%>\%
   filter(PATIENT \%in\% dfStart$PATIENT)
 
 adherences <- pdc_treatment(serialDf = kela, startDates = dfStart, endDates = dfEnd, atcCode = c('^C09', '^C10'), refillPeriod = 90, 
                          idColumn = "PATIENT", dateColumn = "VISIT", atcColumn = "ATC", createGraphs = T, 
                          treatmentBreakDays = c(181,181))
}
}
