% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{plotIndividualScore}
\alias{plotIndividualScore}
\title{Plot the score of an individual patient for all blocks (with binary medication data)}
\usage{
plotIndividualScore(scoreData, patient)
}
\arguments{
\item{scoreData}{data frame containing at least the PATIENT, BLOCK, phenoScore and binaryRAAS}

\item{patient}{patient number to plot}
}
\value{
a ggplot element that can be plotted with the plot function
}
\description{
Plot the score of an individual patient for all blocks (with binary medication data)
}
\keyword{internal}
