% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{Read in a dataset}
\usage{
readData(fileName, descName = NA, doRemove = TRUE, guessRows = NA)
}
\arguments{
\item{fileName}{path to the tab-delimited data file.}

\item{descName}{path to the description file (optional)}

\item{doRemove}{removes PATIENT entries that start with !, DSP, TH or CC}

\item{guessRow}{when no description file is given, how many rows should be used to guess the column type (default: 1000)}
}
\value{
The dataset as a data frame.
}
\description{
Read in a tab-delimited dataset that also has a description file available with column types.
}
\details{
The description file is a direct downlad from the BCOS table descritpions

If only the fileName is given, the function behaves like read_tsv from readr
}
