% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{checkResponseFormat}
\alias{checkResponseFormat}
\title{Checks the format whether the format for the data set is a time block based}
\usage{
checkResponseFormat(df, idColumn, blockColumn, sexColumn, responseColumns)
}
\arguments{
\item{df}{the data frame to check}

\item{idColumn}{name of ID column}

\item{blockColumn}{name of time block column}

\item{sexColumn}{name of the sex column (optional)}

\item{responseColumn}{name of the column containing the patient block value for a biomarker}
}
\value{
NULL if there are format problems
}
\description{
Checks the format whether the format for the data set is a time block based
}
\keyword{internal}
