% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{checkStructureColumn}
\alias{checkStructureColumn}
\title{Checks whether a column is in a data frame and all its rows are not NA}
\usage{
checkStructureColumn(df, columnName)
}
\arguments{
\item{df}{the data frame to check}

\item{columnName}{name of the column}
}
\value{
TRUE if the  column passes the test
}
\description{
Checks whether a column is in a data frame and all its rows are not NA
}
\keyword{internal}
