% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMedicationTimeBlocks.R
\name{createMedicationTimeBlocks}
\alias{createMedicationTimeBlocks}
\title{Calculates the time block data for prescriptions}
\usage{
createMedicationTimeBlocks(serialDf, baselineInfo, studyStartDate, studyEndDate,
  atcCode = "C09", blockDays = 365.25, useDoses = FALSE,
  idColumn = "PATIENT", dateColumn = "VISIT", atcColumn = "ATC",
  dddColumn = "cDDD", dataName = "prescription", longFormat = TRUE,
  noUseAtBaseline = FALSE)
}
\arguments{
\item{serialDf}{the data frame with the serial data}

\item{baselineInfo}{dataframe with ID and baseline date}

\item{studyStartDate}{start Date of the study (has to be of class Date)}

\item{studyEndDate}{end Date of the study (has to be of class Date)}

\item{atcCode}{regular expression for the class of medication ATC code (default: C09 which stands for any RAAS)}

\item{blockDays}{length of a time block in days (default 365.25 days)}

\item{useDoses}{flag indicating whether presence of prescription events or sum of doses should be used in a block (default FALSE)}

\item{idColumn}{name of ID column: default is PATIENT}

\item{dateColumn}{name of date column: default is VISIT. This column has to be of class Date}

\item{atcColumn}{name of the column with the ATC codes: default is ATC}

\item{dataName}{set a user specified name for the data column in the output: default is prescription}

\item{doseColumn}{name of the column with the dosage information: default is cDDD}
}
\value{
time block data and dates in wide format
}
\description{
Calculates the time block data for prescriptions
}
