% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{findClosestToBaseline}
\alias{findClosestToBaseline}
\title{Finds the measurement date that is closest to baseline}
\usage{
findClosestToBaseline(
  serialDf,
  baselineDates,
  blDayDiff = 90,
  idColumn = "PATIENT",
  dateColumn = "VISIT",
  dataColumn = "MEASURE"
)
}
\arguments{
\item{idColumn}{name of ID column: default is PATIENT}

\item{dateColumn}{name of date column: default is VISIT. This column has to be of class Date}

\item{dataColumn}{name of the data column: default is MEASURE}

\item{df}{the data frame to check}

\item{dayDiff}{time around baseline (default: 90 days, will look three months before and after baseline)}
}
\value{
data frame with baseline measure or NA if no measurement was close to baseline
}
\description{
Finds the measurement date that is closest to baseline in the given time period around baseline.
}
\examples{
\dontrun{
baselinePatients <- read_tsv('../data/ourBaselinePatients.txt', guess_max = 3582) \%>\%
    select(PATIENT, VISIT = VISIT_TL)
bmi <- read_tsv('../data/serial_bmi.txt') # contains serial and baseline data
baselinePatients <- findClosestToBaseline(bmi, baselinePatients, blDayDiff = 90)
}
}
